# Conflicts:
#	Assets/Game/PlayerAndCards/PlayerScripts/Player.cs
#	Assets/Scenes/MainScene.unity

+++ Assets/Game/PlayerAndCards/PlayerScripts/Interfaces/IStunnable.cs
@@ -2,7 +2,7 @@
 {
     public interface IStunnable
     {
-        void Stun(float duration);
+        void Stun(int duration);
         bool IsStunned { get; }
     }
 }
